+++ lib/Model/CreateModel.php
@@ -62,7 +62,7 @@ class CreateModel implements ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'id' => 'int32'
+        'id' => 'int64'
     ];
 
     public static function swaggerTypes()
