+++ lib/Model/RemainingCreditModelChild.php
@@ -64,8 +64,8 @@ class RemainingCreditModelChild implements ArrayAccess
       * @var string[]
       */
     protected static $swaggerFormats = [
-        'sms' => 'int32',
-        'email' => 'int32'
+        'sms' => 'int64',
+        'email' => 'int64'
     ];
 
     public static function swaggerTypes()
